updated todos cmps and separated
